Add files via upload
A simple way to save txt files with JavaScript. One limitation is that it can't preserve line breaks ('\n'), that would requires something called base64 encoding, which only supports ASCII letters. If the user is using internet explorer 10 or 11, this should save a txt file with a weird name, but with the same content.